add example/memory

+++ lib/memory/inc/sdf_memory.h
@@ -1,5 +1,5 @@
 #define EMALLOC(n, ppdata) emalloc((n)*sizeof(**(ppdata)), (void**)(ppdata))
 #define EFREE(pdata) efree(pdata)
 
-void emalloc(size_t, void**);
+void emalloc(int, void**);
 void efree(void *ptr);
