+++ lib/memory/src/main.c
@@ -2,11 +2,11 @@
 #include <stdlib.h>
 #include "sdf_memory.h"
 
-void emalloc(size_t size, /**/ void **pq) {
+void emalloc(int size, /**/ void **pq) {
     void *q;
     q = malloc(size);
     if (q == NULL) {
-        fprintf(stderr, "emalloc failed (size=%ld\n", size);
+        fprintf(stderr, "emalloc failed (size=%d\n", size);
         exit(2);
     }
     *pq = q;
